requirements.txt
ray
pandas
pyarrow
runpod~=1.7.7
huggingface-hub
packaging
typing-extensions>=4.8.0
pydantic
pydantic-settings
hf-transfer
transformers>=4.55.0
bitsandbytes>=0.45.0
kernels
torch==2.6.0
pillow==9.5.0
torchvision>=0.18.0
diffusers>=0.31.0
accelerate>=0.33.0
safetensors>=0.4.3
requests>=2.32.0


